エラー分類・ジョブ進捗・DACL 既定値・clippy/CI を改善
- EngineError に taxonomy と Display を追加。76 箇所の `Cuda(String)` catch-all を
  InvalidInput / Unsupported / Internal / OutOfVram へ再分類し、archive framing の
  parse error や base64 の検証 error が「CUDA エラー」を名乗る状態を解消する。
  WinFsp 層の NTSTATUS 対応も変種ごとに割り当て直した。

- job registry に JobProgress（done_bytes / total_bytes）を追加し、status.json へ
  出力する。GUI は経過秒数だけでなく実際の割合と ETA を表示できる。

- hash job を GPU launch budget（約 0.1 秒ぶん）単位のバッチに分割し、バッチごとに
  engine lock を取り直す。従来はツリー全体の hash が終わるまでロックを保持し、
  その間ボリュームへの全 callback が停止していた。CPU path は window ごとに
  進捗を報告する。

- ボリュームの既定 DACL を、マウントしたユーザーの SID + SYSTEM + Administrators に
  限定する。従来は Everyone フルアクセスで、複数アカウントが同時にサインインする
  環境ではディスクの内容が他の全アカウントから読めた。token の照会に失敗した
  場合のみ従来の DACL へ fallback し、その旨を警告する。

- clippy の指摘を全て解消（deny 相当の never_loop error を含む）し、
  `cargo clippy --all-targets -- -D warnings` を CI の lint job として追加。

+++ src/chunk.rs
@@ -31,7 +31,7 @@ impl ChunkAllocator {
         };
         // Mark padding bits (beyond `total`) in the last word as used so they
         // are never handed out.
-        let rem = (total % 64) as u32;
+        let rem = total % 64;
         if rem != 0 {
             let last = words - 1;
             let valid_mask = (1u64 << rem) - 1;
